requirements.txt
django
django-allauth
Pillow
python-dotenv
torch
validators